test: createUser 메서드를 user 패키지의 UserTestUtils로 이동
메서드의 역할에 맞추어 클래스를 변경하였습니다.
