+++ src/test/java/com/potato_y/where_are_you/user/UserTestUtils.java
@@ -1,0 +1,17 @@
+package com.potato_y.where_are_you.user;
+
+import com.potato_y.where_are_you.authentication.domain.oauth.OAuthProvider;
+import com.potato_y.where_are_you.user.domain.User;
+
+public class UserTestUtils {
+
+  public static User createUser(String email, String nickname, String serviceId) {
+    return User.builder()
+        .email(email)
+        .nickname(nickname)
+        .serviceId(serviceId)
+        .password("test_password")
+        .oAuthProvider(OAuthProvider.KAKAO)
+        .build();
+  }
+}
